requirements.txt
streamlit
langchain
huggingface_hub==0.17.1
sentence-transformers>=2.7.0
faiss-cpu
PyPDF2
python-dotenv
langchain-community